refactor(entity): 将 Post 实体类中的 id 类型从 Long 改为 Integer
- 修改了 Post 类中的 id 和 userId 字段类型，从 Long 改为 Integer
- 这个改动可以减少存储空间的使用，并且在大多数情况下，Integer 已经足够表示用户 ID 和帖子 ID

+++ backend/src/main/java/cn/hmck/entity/Post.java
@@ -11,9 +11,9 @@ import java.time.LocalDateTime;
 @TableName("er_post")
 public class Post implements Serializable {
     @TableId(type = IdType.AUTO)
-    private Long id;  // 动态ID
+    private Integer id;  // 动态ID
 
-    private Long userId;  // 发布动态的用户ID
+    private Integer userId;  // 发布动态的用户ID
 
     private String content;  // 动态内容
